requirements.txt
llama-parse==0.6.94
llama-cloud==0.1.46
llama-cloud-services==0.6.94
pydantic==2.12.5
pydantic-settings==2.13.1
langchain==1.2.11
langchain-text-splitters
langchain-google-genai
langchain-chroma
langchain-community
langchain-classic
langchain-cohere
langchain-groq
cohere
chromadb
google-generativeai
python-dotenv
rank_bm25
nest_asyncio
streamlit
langsmith
audio-recorder-streamlit
SpeechRecognition
gTTS
pandas
fpdf2
markdown
fastapi
uvicorn
